requirements.txt
mysql-connector-python
python-dotenv
flask
gunicorn
